fix serializer for array type bug

+++ seatunnel-engine/seatunnel-engine-storage/imap-storage-file/src/main/java/org/apache/seatunnel/engine/imap/storage/file/IMapFileStorage.java
@@ -119,6 +119,7 @@ public class IMapFileStorage implements IMapStorage {
      */
     @Override
     public void initialize(Map<String, Object> configuration) {
+        System.setProperty("protostuff.runtime.preserve_null_elements", "true");
         checkInitStorageProperties(configuration);
         Configuration hadoopConf = (Configuration) configuration.get(HDFS_CONFIG_KEY);
         this.conf = hadoopConf;
@@ -131,6 +132,7 @@ public class IMapFileStorage implements IMapStorage {
         this.businessRootPath = namespace + DEFAULT_IMAP_FILE_PATH_SPLIT + clusterName + DEFAULT_IMAP_FILE_PATH_SPLIT + businessName + DEFAULT_IMAP_FILE_PATH_SPLIT;
         try {
             this.fs = FileSystem.get(hadoopConf);
+            fs.setWriteChecksum(false);
         } catch (IOException e) {
             throw new IMapStorageException("Failed to get file system", e);
         }
@@ -221,7 +223,7 @@ public class IMapFileStorage implements IMapStorage {
     }
 
     @Override
-    public void destroy() {
+    public void destroy(boolean deleteAllFileFlag) {
         log.info("start destroy IMapFileStorage, businessName is {}, cluster name is {}", businessName, region);
         /**
          * 1. close current disruptor
@@ -233,15 +235,16 @@ public class IMapFileStorage implements IMapStorage {
         } catch (IOException e) {
             log.error("close walDisruptor error", e);
         }
-        // delete all files
-        String parentPath = businessRootPath;
+        if (deleteAllFileFlag) {
+            // delete all files
+            String parentPath = businessRootPath;
 
-        try {
-            fs.delete(new Path(parentPath), true);
-        } catch (IOException e) {
-            log.error("destroy IMapFileStorage error,businessName is {}, cluster name is {}", businessName, region, e);
+            try {
+                fs.delete(new Path(parentPath), true);
+            } catch (IOException e) {
+                log.error("destroy IMapFileStorage error,businessName is {}, cluster name is {}", businessName, region, e);
+            }
         }
-
     }
 
     private IMapFileData parseToIMapFileData(Object key, Object value) throws IOException {
